BaSP-TG-83: Fixed post feature and deleted comments in index and AdminForm
