+++ src/Components/Admins/index.js
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import styles from './admins.module.css';
 import List from './List';
-import AddItem from './AdminForm/AdminForm';
 
 const Admins = () => {
-  let [showAddAdmin, setShowAddAdmin] = useState(false);
   const [list, setList] = useState([]);
 
   useEffect(async () => {
@@ -12,7 +10,6 @@ const Admins = () => {
       const response = await fetch(`https://coco-trackgenix-server.vercel.app/admins`);
       const data = await response.json();
       setList(data.data);
-      // console.log('data inicial:', data.data);
     } catch (error) {
       console.error(error);
     }
@@ -20,10 +17,9 @@ const Admins = () => {
 
   const deleteItem = async (_id) => {
     try {
-      const response = await fetch(`https://coco-trackgenix-server.vercel.app/admins/${_id}`, {
+      await fetch(`https://coco-trackgenix-server.vercel.app/admins/${_id}`, {
         method: 'DELETE'
       });
-      console.log('delete response:', response);
       alert('Admin deleted');
     } catch (error) {
       console.error(error);
@@ -31,50 +27,17 @@ const Admins = () => {
     setList(list.filter((listItem) => listItem._id !== _id));
   };
 
-  const addItem = async (e) => {
-    try {
-      // e.preventDefault();
-      const response = await fetch(`https://coco-trackgenix-server.vercel.app/admins`, {
-        method: 'POST',
-        headers: {
-          Accept: 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(e)
-      });
-      alert('Admin created');
-      const data = await response.json();
-      console.log(data);
-      setList([...list, data]);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  const switcher = () => {
-    setShowAddAdmin(showAddAdmin ? (showAddAdmin = false) : (showAddAdmin = true));
-  };
-
-  if (showAddAdmin) {
-    return (
-      <section className={styles.container}>
-        <h2>Admins</h2>
-        <AddItem addItem={addItem} switcher={switcher} />
-      </section>
-    );
-  } else {
-    return (
-      <section className={styles.container}>
-        <h2 className={styles.h2}>Admins</h2>
-        <div>
-          <a href="/admins/add" className={styles.addBtn}>
-            + Add an admin
-          </a>
-          <List list={list} setList={setList} deleteItem={deleteItem} />
-        </div>
-      </section>
-    );
-  }
+  return (
+    <section className={styles.container}>
+      <h2 className={styles.h2}>Admins</h2>
+      <div>
+        <a href="/admins/add" className={styles.addBtn}>
+          + Add an admin
+        </a>
+        <List list={list} setList={setList} deleteItem={deleteItem} />
+      </div>
+    </section>
+  );
 };
 
 export default Admins;
